refactor(waiting): auth 패키지 정리
