+++ src/main/java/auth/LoginController.java
@@ -1,10 +1,10 @@
 package auth;
 
+import auth.dto.TokenRequest;
+import auth.dto.TokenResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
