+++ src/main/java/auth/dto/UserDetails.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.dto;
 
 public class UserDetails {
     private Long id;
