+++ src/main/java/auth/utils/AdminInterceptor.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.utils;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.servlet.HandlerInterceptor;
