+++ src/main/java/auth/utils/AuthenticationException.java
@@ -1,4 +1,4 @@
-package auth;
+package auth.utils;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
