+++ src/main/java/auth/utils/AuthenticationProvider.java
@@ -1,4 +1,7 @@
-package auth;
+package auth.utils;
+
+import auth.dto.TokenRequest;
+import auth.dto.UserDetails;
 
 public interface AuthenticationProvider {
