+++ src/main/java/auth/utils/LoginMemberArgumentResolver.java
@@ -1,5 +1,7 @@
-package auth;
+package auth.utils;
 
+import auth.LoginService;
+import auth.dto.LoginMember;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.support.WebDataBinderFactory;
