+++ src/test/java/nextstep/AbstractE2ETest.java
@@ -1,8 +1,8 @@
 package nextstep;
 
 import io.restassured.RestAssured;
-import auth.TokenRequest;
-import auth.TokenResponse;
+import auth.dto.TokenRequest;
+import auth.dto.TokenResponse;
 import nextstep.member.MemberRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
