+++ src/test/java/nextstep/auth/JwtTokenProviderTest.java
@@ -1,6 +1,6 @@
 package nextstep.auth;
 
-import auth.JwtTokenProvider;
+import auth.utils.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
